[ENGG-3280] fix: tab unsaved changes indication (#2847)
* fix: preview mode when changes

* fix: unsaved changes indication

* fix: unsaved changes indications

* fix: block tab close when unsaved changes
