+++ app/src/componentsV2/Tabs/components/TabItem.tsx
@@ -14,6 +14,16 @@ export const TabItem: React.FC<React.PropsWithChildren<{ store: StoreApi<TabStat
           props.store.getState().setTitle(title);
           incrementVersion();
         },
+
+        setPreview: (preview: boolean) => {
+          props.store.getState().setPreview(preview);
+          incrementVersion();
+        },
+
+        setSaved: (saved: boolean) => {
+          props.store.getState().setSaved(saved);
+          incrementVersion();
+        },
       }}
     >
       {props.children}
