+++ app/src/features/apiClient/screens/apiClient/components/clientView/components/Collection/components/CollectionAuthorizationView/CollectionAuthorizationView.tsx
@@ -1,12 +1,12 @@
 import React, { useCallback, useEffect, useMemo, useState } from "react";
 import { useHasUnsavedChanges } from "hooks";
-import { useTabsLayoutContext } from "layouts/TabsLayout";
 import AuthorizationView from "../../../request/components/AuthorizationView";
 import useEnvironmentManager from "backend/environment/hooks/useEnvironmentManager";
 import { RQAPI } from "features/apiClient/types";
 import { RQButton } from "lib/design-system-v2/components";
 import { KEYBOARD_SHORTCUTS } from "../../../../../../../../../../../src/constants/keyboardShortcuts";
 import { RoleBasedComponent } from "features/rbac";
+import { useGenericState } from "hooks/useGenericState";
 
 interface Props {
   collectionId: string;
@@ -31,17 +31,20 @@ const CollectionAuthorizationView: React.FC<Props> = ({
 }) => {
   const [authOptionsState, setAuthOptionsState] = useState<RQAPI.Auth>(authOptions);
   const [isSaving, setIsSaving] = useState(false);
+  const { setPreview, setSaved } = useGenericState();
 
   const { getVariablesWithPrecedence } = useEnvironmentManager();
   const variables = useMemo(() => getVariablesWithPrecedence(collectionId), [collectionId, getVariablesWithPrecedence]);
 
-  const { updateTab } = useTabsLayoutContext();
-
   const { hasUnsavedChanges, resetChanges } = useHasUnsavedChanges(authOptionsState);
 
   useEffect(() => {
-    updateTab(collectionId, { hasUnsavedChanges: hasUnsavedChanges });
-  }, [updateTab, collectionId, hasUnsavedChanges]);
+    setSaved(hasUnsavedChanges);
+
+    if (hasUnsavedChanges) {
+      setPreview(false);
+    }
+  }, [setSaved, setPreview, hasUnsavedChanges]);
 
   const onSaveAuthData = useCallback(() => {
     setIsSaving(true);
